DLPX-31550 dx.core.data shouldn't be exposing null
dxCoreData model.js was not being highly consistent about when it was returning null versus undefined to clients.  The server, as it communicates with JSON, necessarily delivers nulls, but undefined is ultimately more "natural" in JavaScript.  So, this change makes the system consistent: model.js will only deliver undefined values to its clients, and will deliver only null to the server.  (it will accept either null or undefined as inputs, however)