DLPX-29226 dx.core.data won't allow a type to upgrade or downgrade
PROBLEM
===
While working on a new feature, I found update was failing.

The cause was that this involved doing an update that was changing an embedded object of type Credentials to PasswordCredentials.  The jsonIzeForUpdate routine in dxCoreData was not prepared to be doing this (that is to say, it wasn't ready to be doing type conversions.

This is a blocker for this project.  And the gotcha is that as soon as update would work, dxSet in model.js is going to encounter DLPX-29226 since this update is going to involve a change in the data type of the embedded object when the notification system learns of it. So, while there was an easy workaround for this particular problem in dxSet, it seemed like it really is time to deal with the longstanding DLPX-29226 bug.

CAUSE
===
When dxCore was created, the only case forseen needing type converstion was as a model was moving from pre-ready to ready.  Talk about an oversight!  Clearly, ClientModels could change rather considerably.  And this case of needing to do an update causing problems has never come up yet.

FIX
====
There are two fixes here:

* jsonizeForUpdate has been changed to cope with an update asking to change the type of an embedded model. This is a near trivial fix and just involves inserting a newly allocated object into the process to allow this to work smoothly.
* dxSet needs some notable changes to allow for types to change.  You can think of this in two ways: A top level object might change while it is being retrieved from the server (e.g. at getServerModel('IngresContainer', 'Container') would give you a Container which will be changed to IngresContainer later.  Alternately, you might have an embedded model which could go from type ParentType to type ChildType, from Child to Parent, or even from Cousin to Sibling.

This does *not* fix the dx.core.ko system on top of dx.core.data.  That's a completely different can of worms. I'll be filing a bug for that so it can be dealt with eventually, now that the underpinnings are present.

The second bullet deserves more commentary:

1. For the case of a "top level" object, we only allow type specialization.  This is "arbitrary", in that we could allow any kind of type transformation (e.g. var foo = newClientModel('Container'); foo.set('type', 'User');).  For now, I don't see a compelling reason to allow that. It's just a couple lines to change it should we ever want to.
2. Changing the embedded models is reasonably straightforward as well. I generalized the convertToSubtype routine to be a converToType routine, which will now convert between any two types. yay. In particular, I now pass a special option flag to dxSet internally when setting the embedded models so that all the type conversion goes through one place.  Note that all the setting attempts to preserve particular object instances as much as possible, so event listeners set on one model will be preserved even across a type change.
3. The nightmare of this change turns out to be handling the attribute change events. In fact, if you think about it carefully, converting a client model to a subtype (which was allowed before) wasn't sending the right events (maybe an interesting debugging interview question?)  So, there's a bunch of new code for these events.  What we now do is when we notice we're doing a type conversion, we set a flag and start capturing snapshots of the attributes of the model. We get a snapshot before the conversion, after the conversion, and then implicitly the set of attributes after the main Backbone.set() operation is done.  We then do various set comparisons of these properties to determine: which properties were removed, which were added, which were changed, and dispatch the appropriate events.
As an aside here, I'll mention: Backbone is really disciplined about it's events. change:attrName events are always all sent before change events. Furthermore, all events are sent only once all attributes have been set, so setting is an atomic operation from the viewpoint of an observer.  It was important that the same behavior is preserved here, so while we could be doing all the event dispatching in places closer to where the changes happen, this actually waits until the end of the routine to start determining which events to dispatch.
The ugliest part of this is that because we're sitting cleanly on top of Backbone, we need to do a couple hinky things with it.  If we do a type conversion, it is quite possible that the Backbone.set() routine will dispatch a change event, because it will think it's done all the setting. However, we may have events to dispatch after that. Further, there's one corner case where Backbone will think a attribute has changed and so dispatch a change:attrName event, but it is wrong because its set is returning the attribute to the same value it was before the conversion.  To deal with this, we will briefly intercept all calls to trigger() and then later play them back, rearranging the change event so it comes at the end, and suppressing that corner case.