DLPX-37448 When tests fail with things in the cache, the reports should be less noisy
PROBLEM
While working with Manoj today on a problem, we had a case where reams of tests were failing, and spewing out enormous amounts of cruft.  While this problem has grated on me when I'm doing work, it was doubly grating to have someone outside of the group having to deal with it.

FIX
So, while I was repairing some unit tests, I changed how we're reporting this info. This gives, in some cases, like a 100% reduction in output.

Note: the goal remains the same: give you enough information to help identify what is in the cache without your needing to bring up the debugger.  Collections show you what is in them, models show you the list of event callbacks. The difference is now we extract the model id's and just show those for the collections, and we extract the function names (when they aren't anonymous) on the event callbacks and list those.

WHILE I WAS HERE
I have also broken up the mode.unit.js file into four (!) files so that we no longer get errors from babel that it can't optmize the file layout.