Fixed longstanding annoying problem with two sessions in one browser.
For a long time, I've seen irregular, and seemingly random, behavior when
running two sessions in one browser at the same time. It dawned on me that
I was not attening to the "channel" value in the notification requests, and
so two tabs were sharing the same notification queue.
Fixed this by altering the session to keep a set of channels to deliver
notifications through.